Hopefully last fix

+++ src/VectorBackedLists.jl
@@ -70,9 +70,9 @@ state of the underlying container.
 function list(data)
     n = length(data)
     nodes = Vector{Node}(undef, n+2)
-    nodes[1] = Node(0,2,0)
+    nodes[1] = Node(UInt32(0), UInt32(2), UInt32(0))
     for i in 2:n+1; nodes[i] = Node(UInt(i-1), UInt32(i+1), UInt32(i-1)); end
-    nodes[end] = Node(0,0,n+1)
+    nodes[end] = Node(UInt32(0), UInt32(0), UInt32(n+1))
     VectorBackedList{eltype(data), typeof(data)}(data, nodes, UInt32(1), UInt32(n+2))
 end
